extract code from sigma_protocol.rs into sub-modules;
